Hide song downloader implementation
